refactor: governance section
