+++ src/pages/governance/new.tsx
@@ -1,22 +1,18 @@
 // sections
-// import { GovernanceListView } from '../../../sections/governance/view';
-import ComingSoonView from '@src/sections/coming-soon/view.tsx';
-import BlankView from '@src/sections/blank/view.tsx';
+import { GovernanceCreateView } from '@src/sections/governance';
 import { GLOBAL_CONSTANTS } from '@src/config-global.ts';
 import { OgMetaTags } from '@src/components/og-meta-tags.tsx';
 
 // ----------------------------------------------------------------------
 
-export default function GovernanceList() {
+export default function UserCreatePage() {
   return (
     <OgMetaTags
       title="Watchit: Governance (COMING SOON)"
       description="Discover the latest decentralized creations on Watchit. Powered by Web3 & AI."
       url={`${GLOBAL_CONSTANTS.BASE_URL}/governance`}
     >
-      <BlankView>
-        <ComingSoonView />
-      </BlankView>
+      <GovernanceCreateView />
     </OgMetaTags>
   );
 }
