+++ src/sections/governance/index.tsx
@@ -1,0 +1,3 @@
+export { default as GovernanceListView } from '@src/sections/governance/views/governance-list-view';
+export { default as GovernanceCreateView } from '@src/sections/governance/views/governance-create-view';
+export { default as GovernanceDetailsView } from '@src/sections/governance/views/governance-details-view';
