--- src/sections/governance/view/index.ts
@@ -1,3 +1,0 @@
-export { default as GovernanceListView } from './governance-list-view';
-export { default as GovernanceCreateView } from './governance-create-view';
-export { default as GovernanceDetailsView } from './governance-details-view';
